implemented service communication and underlying valueobjects for usage mostly in GUI building and binding

+++ ark-net/io/ark/model/TransactionVO.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -11,8 +12,8 @@ namespace io.ark.model
         public string id { get; set; }
         public int type { get; set; }
         public int timestamp { get; set; }
-        public int amount { get; set; }
-        public int fee { get; set; }
+        public BigInteger amount { get; set; }
+        public BigInteger fee { get; set; }
         public string senderId { get; set; }
         public string recipientId { get; set; }
         public string senderPublicKey { get; set; }
